// NSDL_FVCI_UAT_RELEASE/03-03-2025/Website Deployment/src/swagger/model/draftFvciGlobalCustodianDetailsDto.ts
/**
 * NSDL API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface DraftFvciGlobalCustodianDetailsDto { 
    id?: number;
    fvciApplicationId?: string;
    name?: string;
    country?: string;
    registrationNumber?: string;
    address?: string;
    createdAt?: Date;
    updatedAt?: Date;
}